refactor(widgets): Extract layout helpers to widgets/_helpers.py
Move all widget-only helpers to widgets/_helpers.py so widget modules
import from a sibling instead of reaching back into the rendering
pipeline.
