+++ custom_components/eink_dashboard/widgets/entities.py
@@ -14,13 +14,13 @@ from ..const import (
     DEFAULT_ROW_H,
     PADDING,
 )
-from ..svg_render import (
+from ..svg_render import _mdi_svg_filter
+from ._helpers import (
     _ACTIVE_STATES,
     _auto_row_height,
     _card_insets,
     _color_context,
     _fmt,
-    _mdi_svg_filter,
     _metrics_context,
     _title_layout,
     _widget_dim,
