+++ custom_components/eink_dashboard/widgets/heading.py
@@ -14,11 +14,11 @@ from ..const import (
     DEFAULT_ROW_H,
     PADDING,
 )
-from ..svg_render import (
+from ..svg_render import _mdi_svg_filter
+from ._helpers import (
     _card_insets,
     _color_context,
     _fmt,
-    _mdi_svg_filter,
     _metrics_context,
     _widget_dim,
 )
