+++ custom_components/eink_dashboard/widgets/separator.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Any
 
 from ..const import COLOR_BLACK, COLOR_GRAY, PADDING
-from ..svg_render import _color_context, _widget_dim
+from ._helpers import _color_context, _widget_dim
 
 type Widget = dict[str, Any]
 type DisplayConfig = dict[str, Any]
