+++ custom_components/eink_dashboard/widgets/weather.py
@@ -15,12 +15,14 @@ from ..const import (
 from ..svg_render import (
     _ICONS_DIR,
     _build_inline_svg,
+    _load_svg_paths,
+    _weather_svg_filter,
+)
+from ._helpers import (
     _card_insets,
     _color_context,
     _fmt,
-    _load_svg_paths,
     _metrics_context,
-    _weather_svg_filter,
     _widget_dim,
 )
